pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "elysia-ai"
version = "0.3.dev1"
description = "Elysia is an open-source agentic platform for searching data. It is built with customisation in mind, allowing you to build agents and tools that are tailored to your specific use case. It uses Weaviate as the default retrieval tools, and can interface with your data stored in a Weaviate cluster."
readme = "README.md"
authors = [
    {name = "Danny", email = "danny@weaviate.io"}
]
requires-python = ">=3.10.0,<3.13.0"
classifiers = [
    "Development Status :: 4 - Beta"
]
keywords=["ai", "agentic", "weaviate", "retrieval", "search", "llm"]
dependencies = [
    "apscheduler==3.11.0",
    "bcrypt>=4.3.0",
    "dspy-ai>=3.0.3",
    "fastapi[standard]>=0.115.11",
    "httpx==0.28.1",
    "pympler==1.1",
    "python-multipart==0.0.18",
    "rich>=13.7.1,<=14.0.0",
    "blis>=1.2.0,<1.3.0",
    "spacy==3.8.7",
    "uvicorn[standard]==0.35.0",
    "weaviate-client>=4.16.7",
    "nest_asyncio==1.6.0",
    "psutil==7.0.0",
    "cryptography>=44.0.3",
    "matplotlib>=3.10.3",
    "litellm>=1.74.15,<1.76.0",
    "thinc>=8.3.4,<8.4.0",
    "pip>=25.2",
    "PyArabic==0.6.15"
]
packages=[{include="elysia"}]

[project.urls]
"Homepage" = "https://elysia.weaviate.io"
"Documentation" = "https://weaviate.github.io/elysia/"

[project.optional-dependencies]
dev = [
    "accelerate==1.5.2",
    "deepeval>=3.2.6",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.21.2",
    "setuptools==78.1.0",
    "mkdocs-material[imaging]==9.6.12",
    "mkdocstrings[python]==0.29.1",
    "pillow==10.4.0",
    "cairosvg==2.7.1",
    "websocket-client==1.8.0",
    "pytest-cov>=6.2.1"
]

[project.scripts]
elysia = "elysia.api.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["elysia"]
